pyproject.toml
[project]
name = "blog-post"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.6",
    "djangorestframework>=3.16.1",
    "dotenv>=0.9.9",
    "django-cors-headers>=4.5.0",
]
